lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
ask-sdk-dynamodb-persistence-adapter==1.11.0
pandas==0.25.2
requests